updated periphery with Base support for MultiInvoker and Base coverage
