+++ packages/periphery/hardhat.config.ts
@@ -1,16 +1,5 @@
 import defaultConfig, { OPTIMIZER_ENABLED, SOLIDITY_VERSION } from '../common/hardhat.default.config'
 
-const controllerOverrides = {
-  version: SOLIDITY_VERSION,
-  settings: {
-    optimizer: {
-      enabled: OPTIMIZER_ENABLED,
-      runs: 80000,
-    },
-    viaIR: OPTIMIZER_ENABLED,
-  },
-}
-
 const multiInvokerOverrides = {
   version: SOLIDITY_VERSION,
   settings: {
@@ -23,7 +12,6 @@ const multiInvokerOverrides = {
 }
 
 export const solidityOverrides = {
-  'contracts/Controller_Arbitrum.sol': controllerOverrides,
   'contracts/MultiInvoker.sol': multiInvokerOverrides,
 }
 const config = defaultConfig({
