+++ packages/periphery/package.json
@@ -18,15 +18,13 @@
     "gasReport": "REPORT_GAS=true OPTIMIZER_ENABLED=true yarn test:integration",
     "test": "HARDHAT_SHOW_STACK_TRACES=true hardhat test test/unit/**/*",
     "test:integration": "FORK_ENABLED=true FORK_BLOCK_NUMBER=17433155 hardhat test test/integration/**/*",
-    "test:integrationArbitrum": "FORK_ENABLED=true FORK_NETWORK=arbitrum FORK_BLOCK_NUMBER=243648015 hardhat test test/integration/**/*",
     "test:integrationBase": "FORK_ENABLED=true FORK_NETWORK=base FORK_BLOCK_NUMBER=21067741 hardhat test test/integration/**/*",
     "coverage": "hardhat coverage --testfiles 'test/unit/**/*'",
     "coverage:integration": "FORK_ENABLED=true FORK_BLOCK_NUMBER=17433155 hardhat coverage --testfiles 'test/integration/**/*'",
-    "coverage:integrationArbitrum": "FORK_ENABLED=true FORK_NETWORK=arbitrum FORK_BLOCK_NUMBER=243648015 hardhat coverage --testfiles 'test/integration/**/*'",
+    "coverage:integrationBase": "FORK_ENABLED=true FORK_NETWORK=base FORK_BLOCK_NUMBER=21067741 hardhat coverage --testfiles 'test/integration/**/*'",
     "lint": "eslint --fix --ext '.ts,.js' ./ && solhint 'contracts/**/*.sol' --fix",
     "format": "prettier -w .",
     "clean": "rm -rf cache artifacts types/generated deployments/localhost",
-    "node:fork:arbitrumGoerli": "FORK_ENABLED=true FORK_NETWORK=arbitrumGoerli NODE_INTERVAL_MINING=500 hardhat node",
     "prepack": "yarn clean && yarn build && find artifacts/contracts -name '*.dbg.json' -type f -delete"
   },
   "author": "",
