+++ packages/periphery/test/helpers/mainnetHelpers.ts
@@ -10,7 +10,7 @@ export const DSU_ADDRESS = '0x605D26FBd5be761089281d5cec2Ce86eeA667109' // Digit
 export const DSU_BATCHER = '0xAEf566ca7E84d1E736f999765a804687f39D9094'
 export const DSU_RESERVE = '0xD05aCe63789cCb35B9cE71d01e4d632a0486Da4B'
 
-export const USDC_ADDRESS = '0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48' // Arbitrum native USDC (not USDC.e), a 6-decimal token
+export const USDC_ADDRESS = '0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48' // USDC, a 6-decimal token
 export const USDC_HOLDER = '0x0A59649758aa4d66E25f08Dd01271e891fe52199' // Maker PSM has 520mm USDC at height 17433155
 
 export async function fundWalletDSU(wallet: SignerWithAddress, amount: BigNumber): Promise<void> {
