+++ packages/periphery/test/unit/TriggerOrders/Manager.test.ts
@@ -11,21 +11,17 @@ import { IERC20, IFactory, IMarketFactory, IMarket, IOracleProvider } from '@per
 
 import {
   AggregatorV3Interface,
-  ArbGasInfo,
-  IAccount,
-  IAccount__factory,
-  IController,
   IEmptySetReserve,
   IMargin,
   IOrderVerifier,
-  Manager_Arbitrum,
-  Manager_Arbitrum__factory,
+  Manager_Optimism,
+  Manager_Optimism__factory,
   OrderVerifier__factory,
 } from '../../../types/generated'
 import { signCancelOrderAction, signCommon, signPlaceOrderAction } from '../../helpers/TriggerOrders/eip712'
 import { OracleVersionStruct } from '../../../types/generated/contracts/TriggerOrders/test/TriggerOrderTester'
 import { Compare, compareOrders, DEFAULT_TRIGGER_ORDER, Side } from '../../helpers/TriggerOrders/order'
-import { deployController } from '../../helpers/setupHelpers'
+import { mockGasInfo } from '../../helpers/baseHelpers'
 
 const { ethers } = HRE
 
@@ -55,7 +51,7 @@ describe('Manager', () => {
   let usdc: FakeContract<IERC20>
   let dsu: FakeContract<IERC20>
   let reserve: FakeContract<IEmptySetReserve>
-  let manager: Manager_Arbitrum
+  let manager: Manager_Optimism
   let marketFactory: FakeContract<IMarketFactory>
   let margin: FakeContract<IMargin>
   let market: FakeContract<IMarket>
@@ -94,7 +90,7 @@ describe('Manager', () => {
     market.margin.returns(margin.address)
 
     // deploy the order manager
-    manager = await new Manager_Arbitrum__factory(owner).deploy(
+    manager = await new Manager_Optimism__factory(owner).deploy(
       usdc.address,
       dsu.address,
       reserve.address,
@@ -133,10 +129,7 @@ describe('Manager', () => {
 
   before(async () => {
     ;[owner, userA, userB, keeper] = await ethers.getSigners()
-    // Hardhat testnet does not support Arbitrum built-ins; need this for realistic keeper fees
-    await smock.fake<ArbGasInfo>('ArbGasInfo', {
-      address: '0x000000000000000000000000000000000000006C',
-    })
+    await mockGasInfo()
   })
 
   beforeEach(async () => {
